ignore https as false postive for slackwebhook detector
