+++ .gitignore
@@ -9,3 +9,4 @@ tmp/go-test.json
 .captain/detectors/timings.yaml
 .captain/detectors/quarantines.yaml
 .captain/detectors/flakes.yaml
+.vscode
